feat(api): add academic template
