+++ api/pdf-templates/academic-template/coverpage-template.html
@@ -1,0 +1,138 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Cyber Attacks on Federated Learning</title>
+    <style>
+        body {
+            font-family: "Times New Roman", Times, serif;
+            margin: 20px;
+        }
+
+        header {
+            text-align: center;
+            padding: 20px 0;
+        }
+        main {
+            line-height: 1.6;
+            column-count: 2;
+
+        }
+
+        h1 {
+            font-size: 24px;
+        }
+
+        h2 {
+            font-size: 18px;
+        }
+
+        .authors {
+            font-style: italic;
+        }
+
+        .authors span::before {
+            content: '[1]';
+        }
+
+        .lab {
+            font-weight: bold;
+            margin-top: 10px;
+        }
+
+        .lab::before {
+            content: '[1] ';
+        }
+
+        .abstract {
+            margin-top: 20px;
+        }
+
+        .paper-content {
+            margin-top: 30px;
+        }
+        /* Table Styles */
+        table {
+            width: 100%;
+            border-collapse: collapse;
+            margin-top: 20px;
+        }
+
+        th, td {
+            border: 1px solid #ddd;
+            padding: 8px;
+            text-align: left;
+        }
+
+        th {
+            background-color: #f2f2f2;
+        }
+        /* Add your custom styles here */
+
+
+    </style>
+</head>
+<body>
+<header>
+    <h1>Cyber Attacks on Federated Learning</h1>
+    <div class="authors">
+        <span>Wang Lee</span>,
+        <span>Adina Cazalens</span>
+    </div>
+    <div class="lab">Red4Sec</div>
+</header>
+
+<main>
+    <section class="abstract">
+        <h2>Abstract</h2>
+        <p>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed aliquam eleifend urna, eget tristique libero convallis a. Vivamus vel velit libero. Sed ac lectus nec orci pharetra hendrerit. Suspendisse potenti. Nulla facilisi. In ac gravida dui, non tincidunt turpis. Nullam a tortor vel neque interdum pellentesque. Proin sodales bibendum semper. Sed bibendum, erat vel commodo tincidunt, justo lorem consectetur erat, vel facilisis tortor nisi a nunc. Nullam sit amet erat non tortor cursus dapibus.
+        </p>
+    </section>
+
+    <section class="paper-content">
+        <h2>Introduction</h2>
+        <p>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed aliquam eleifend urna, eget tristique libero convallis a. Vivamus vel velit libero. Sed ac lectus nec orci pharetra hendrerit. Suspendisse potenti. Nulla facilisi. In ac gravida dui, non tincidunt turpis. Nullam a tortor vel neque interdum pellentesque. Proin sodales bibendum semper. Sed bibendum, erat vel commodo tincidunt, justo lorem consectetur erat, vel facilisis tortor nisi a nunc. Nullam sit amet erat non tortor cursus dapibus.
+        </p>
+
+        <!-- Add more sections and content here -->
+
+    </section>
+
+    <section class="paper-content">
+        <h2>Introduction</h2>
+        <p>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed aliquam eleifend urna, eget tristique libero convallis a. Vivamus vel velit libero. Sed ac lectus nec orci pharetra hendrerit. Suspendisse potenti. Nulla facilisi. In ac gravida dui, non tincidunt turpis. Nullam a tortor vel neque interdum pellentesque. Proin sodales bibendum semper. Sed bibendum, erat vel commodo tincidunt, justo lorem consectetur erat, vel facilisis tortor nisi a nunc. Nullam sit amet erat non tortor cursus dapibus.
+        </p>
+
+        <!-- Add more sections and content here -->
+
+        <!-- Sample Table -->
+        <table>
+            <thead>
+            <tr>
+                <th>Column 1</th>
+                <th>Column 2</th>
+                <th>Column 3</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr>
+                <td>Data 1</td>
+                <td>Data 2</td>
+                <td>Data 3</td>
+            </tr>
+            <tr>
+                <td>Data 4</td>
+                <td>Data 5</td>
+                <td>Data 6</td>
+            </tr>
+            </tbody>
+        </table>
+    </section>
+</main>
+
+</body>
+</html>
