+++ api/pdf-templates/hackmanit-template/coverpage-template.html
@@ -57,7 +57,7 @@
                     text-align: center;
                     color: gray;
                     font-weight: 300;
-                    font-size: 1em;   
+                    font-size: 1em;
                 }
                 .leader-name {
                     color: red;
