+++ api/pdf-templates/hackmanit-template/weaknesses-template.html
@@ -130,6 +130,7 @@
                 width: 100%;
                 object-fit: cover;
             }
+
         </style>
         <div class="inline-title-logo">
             <img alt="logo-company" id="logo"
